+++ RISCV_final_project/mars/mips/instructions/syscalls/SyscallClose.java
@@ -50,6 +50,6 @@ WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
    * Performs syscall function to close file descriptor given in $a0.
    */
        public void simulate(ProgramStatement statement) throws ProcessingException {
-         SystemIO.closeFile(RV32IRegisters.getValue(4)); 
+         SystemIO.closeFile(RV32IRegisters.getValue(4).intValue());
       }
    }
