+++ RISCV_final_project/mars/mips/instructions/syscalls/SyscallInputDialogDouble.java
@@ -61,16 +61,16 @@ WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 
          String message = new String(); // = "";
-         int byteAddress = RV32IRegisters.getValue(4);
+         int byteAddress = RV32IRegisters.getValue(4).intValue();
          char ch[] = { ' '}; // Need an array to convert to String
          try
          {
-            ch[0] = (char) Globals.memory.getByte(byteAddress);
+            ch[0] = (char) Globals.memory.getByte(byteAddress).byteValue();
             while (ch[0] != 0) // only uses single location ch[0]
             {
                message = message.concat(new String(ch)); // parameter to String constructor is a char[] array
                byteAddress++;
-               ch[0] = (char) Globals.memory.getByte(byteAddress);
+               ch[0] = (char) Globals.memory.getByte(byteAddress).byteValue();
             }
          }
              catch (AddressErrorException e)
